FEATURE: move inflector so plugins can lean on it
Prior to this I don't think plugins had any way to register custom inflections
safely.

On eval ... intializers have not been run yet
after_initialize is already too late
