+++ config/application.rb
@@ -28,6 +28,28 @@ require_relative "../lib/plugin_gem"
 
 # Global config
 require_relative "../app/models/global_setting"
+
+# This custom inflector is needed because of our jobs directory structure.
+# Ideally, we should not prefix our jobs with a `Jobs` namespace but instead
+# have a `Job` suffix to follow the Rails conventions on naming.
+#
+# Based on:
+# https://github.com/rails/rails/blob/75e6c0ac/railties/lib/rails/autoloaders/inflector.rb#L7-L19
+# Additonally this allows plugins to override the inflections cause it is defined prior to plugin init
+module DiscourseInflector
+  @overrides = {}
+
+  def self.camelize(basename, abspath)
+    return basename.camelize if abspath.ends_with?("onceoff.rb")
+    return "Jobs" if abspath.ends_with?("jobs/base.rb")
+    @overrides[basename] || basename.camelize
+  end
+
+  def self.inflect(overrides)
+    @overrides.merge!(overrides)
+  end
+end
+
 GlobalSetting.configure!
 if GlobalSetting.load_plugins?
   # Support for plugins to register custom setting providers. They can do this
